update functionality done

+++ src/Pages/MyRooms/MyRooms.jsx
@@ -32,19 +32,31 @@ const MyRooms = () => {
     console.log(checkCancel);
 
     if (checkCancel) {
-      fetch(`http://localhost:5001/booking/${id}`, {
-        method: "DELETE",
-      })
-        .then((res) => res.json())
-        .then((data) => {
-          console.log(data);
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          fetch(`http://localhost:5001/booking/${id}`, {
+            method: "DELETE",
+          })
+            .then((res) => res.json())
+            .then((data) => {
+              console.log(data);
 
-          if (data.deletedCount > 0) {
-            Swal.fire("Success!", "Booking Cancellation Done", "success");
-            const remaining = myRooms.filter((room) => room._id !== id);
-            setMyRooms(remaining);
-          }
-        });
+              if (data.deletedCount > 0) {
+                Swal.fire("Success!", "Booking Cancellation Done", "success");
+                const remaining = myRooms.filter((room) => room._id !== id);
+                setMyRooms(remaining);
+              }
+            });
+        }
+      });
     } else {
       Swal.fire("Error!", "Booking cancellation time expired", "error");
     }
